fixed bug that showed value field in filter even when it was not needed; also started refactoring redundant code between application and website logic
